[ts-common] error check moved under ts-common
