+++ packages/lib/ts-common/src/index.ts
@@ -5,3 +5,4 @@ export * from './string'
 export * from './logging'
 export * from './async'
 export * from './signals'
+export * from './error'
